+++ db/migrate/20260713000001_add_catalog_model_count_to_completion_kit_provider_credentials.rb
@@ -1,0 +1,5 @@
+class AddCatalogModelCountToCompletionKitProviderCredentials < ActiveRecord::Migration[8.1]
+  def change
+    add_column :completion_kit_provider_credentials, :catalog_model_count, :integer
+  end
+end