+++ standalone/db/migrate/20260714011534_add_catalog_model_count_to_completion_kit_provider_credentials.completion_kit.rb
@@ -1,0 +1,6 @@
+# This migration comes from completion_kit (originally 20260713000001)
+class AddCatalogModelCountToCompletionKitProviderCredentials < ActiveRecord::Migration[8.1]
+  def change
+    add_column :completion_kit_provider_credentials, :catalog_model_count, :integer
+  end
+end